+++ api/src/app.ts
@@ -46,7 +46,7 @@ import { checkIP } from './middleware/check-ip';
 import cors from './middleware/cors';
 import errorHandler from './middleware/error-handler';
 import extractToken from './middleware/extract-token';
-import rateLimiter from './middleware/rate-limiter';
+import rateLimiter from './middleware/rate-limiter-ip';
 import rateLimiterGlobal from './middleware/rate-limiter-global';
 import sanitizeQuery from './middleware/sanitize-query';
 import schema from './middleware/schema';
