+++ api/src/middleware/rate-limiter-ip.ts
@@ -2,7 +2,7 @@ import { RequestHandler } from 'express';
 import ms from 'ms';
 import { RateLimiterMemcache, RateLimiterMemory, RateLimiterRedis } from 'rate-limiter-flexible';
 import env from '../env';
-import { HitRateLimitException } from '../exceptions';
+import { HitRateLimitException } from '../exceptions/index';
 import { createRateLimiter } from '../rate-limiter';
 import asyncHandler from '../utils/async-handler';
 import { getIPFromReq } from '../utils/get-ip-from-req';
