+++ api/src/rate-limiter.ts
@@ -13,42 +13,46 @@ import { getConfigFromEnv } from './utils/get-config-from-env';
 type IRateLimiterOptionsOverrides = Partial<IRateLimiterOptions> | Partial<IRateLimiterStoreOptions>;
 
 export function createRateLimiter(
-	envKey = 'RATE_LIMITER',
+	configPrefix = 'RATE_LIMITER',
 	configOverrides?: IRateLimiterOptionsOverrides
 ): RateLimiterAbstract {
 	switch (env.RATE_LIMITER_STORE) {
 		case 'redis':
-			return new RateLimiterRedis(getConfig('redis', envKey, configOverrides));
+			return new RateLimiterRedis(getConfig('redis', configPrefix, configOverrides));
 		case 'memcache':
-			return new RateLimiterMemcache(getConfig('memcache', envKey, configOverrides));
+			return new RateLimiterMemcache(getConfig('memcache', configPrefix, configOverrides));
 		case 'memory':
 		default:
-			return new RateLimiterMemory(getConfig('memory', envKey, configOverrides));
+			return new RateLimiterMemory(getConfig('memory', configPrefix, configOverrides));
 	}
 }
 
-function getConfig(store: 'memory', envKey: string, overrides?: IRateLimiterOptionsOverrides): IRateLimiterOptions;
+function getConfig(
+	store: 'memory',
+	configPrefix: string,
+	overrides?: IRateLimiterOptionsOverrides
+): IRateLimiterOptions;
 function getConfig(
 	store: 'redis' | 'memcache',
-	envKey: string,
+	configPrefix: string,
 	overrides?: IRateLimiterOptionsOverrides
 ): IRateLimiterStoreOptions;
 function getConfig(
 	store: 'memory' | 'redis' | 'memcache' = 'memory',
-	envKey = 'RATE_LIMITER',
+	configPrefix = 'RATE_LIMITER',
 	overrides?: IRateLimiterOptionsOverrides
 ): IRateLimiterOptions | IRateLimiterStoreOptions {
-	const config: any = getConfigFromEnv(`${envKey}_`, `${envKey}_${store}_`);
+	const config: any = getConfigFromEnv(`${configPrefix}_`, `${configPrefix}_${store}_`);
 
 	if (store === 'redis') {
 		const Redis = require('ioredis');
 		delete config.redis;
-		config.storeClient = new Redis(env[`${envKey}_REDIS`] || getConfigFromEnv(`${envKey}_REDIS_`));
+		config.storeClient = new Redis(env[`${configPrefix}_REDIS`] || getConfigFromEnv(`${configPrefix}_REDIS_`));
 	}
 
 	if (store === 'memcache') {
 		const Memcached = require('memcached');
-		config.storeClient = new Memcached(env[`${envKey}_MEMCACHE`], getConfigFromEnv(`${envKey}_MEMCACHE_`));
+		config.storeClient = new Memcached(env[`${configPrefix}_MEMCACHE`], getConfigFromEnv(`${configPrefix}_MEMCACHE_`));
 	}
 
 	delete config.enabled;
